feat: also modifies multi language input
